requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
astroid==2.15.4
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
dill==0.3.6
flake8==6.0.0
Flask==2.3.2
Flask-RESTful==0.3.9
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gptcache==0.1.21
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.142
lazy-object-proxy==1.9.0
llama-index==0.5.27
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
platformdirs==3.5.0
plumbum==1.8.1
ply==3.11
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pylint==2.17.3
pypandoc==1.11
pypandoc-binary==1.11
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.48
tenacity==8.2.2
tiktoken==0.3.3
tomlkit==0.11.8
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.3.3
wrapt==1.15.0
yarl==1.9.2
